+++ src/services/logging/CoutFrameworkLogger.cpp
@@ -11,7 +11,7 @@ void Ichor::CoutFrameworkLogger::trace(const char *filename_in, int line_in, con
     if(_level <= LogLevel::LOG_TRACE) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
 
@@ -20,7 +20,7 @@ void Ichor::CoutFrameworkLogger::debug(const char *filename_in, int line_in, con
     if(_level <= LogLevel::LOG_DEBUG) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
 
@@ -29,7 +29,7 @@ void Ichor::CoutFrameworkLogger::info(const char *filename_in, int line_in, cons
     if(_level <= LogLevel::LOG_INFO) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
 
@@ -38,7 +38,7 @@ void Ichor::CoutFrameworkLogger::warn(const char *filename_in, int line_in, cons
     if(_level <= LogLevel::LOG_WARN) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
 
@@ -47,7 +47,7 @@ void Ichor::CoutFrameworkLogger::error(const char *filename_in, int line_in, con
     if(_level <= LogLevel::LOG_ERROR) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
